% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{fct_drop}
\alias{fct_drop}
\title{Drop unnused levels}
\usage{
fct_drop(f)
}
\arguments{
\item{f}{A factor}
}
\description{
Same as \code{\link{droplevels}()}, but with consistent name scheme.
}
\examples{
f <- factor(c("a", "b"), levels = c("a", "b", "c"))
f
fct_drop(f)
}

