% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_na.R
\name{fct_explicit_na}
\alias{fct_explicit_na}
\title{Make missing values explicit}
\usage{
fct_explicit_na(f, na_level = "(Missing)")
}
\arguments{
\item{f}{A factor (or character vector).}

\item{na_level}{Level to use for missing values: this is what \code{NA}s will be
changed to.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because the terminology is confusing;
please use \code{\link[=fct_na_value_to_level]{fct_na_value_to_level()}} instead.

This gives missing values an explicit factor level, ensuring that they
appear in summaries and on plots.
}
\examples{
f1 <- factor(c("a", "a", NA, NA, "a", "b", NA, "c", "a", "c", "b"))
fct_count(f1)
table(f1)
sum(is.na(f1))

# previously
f2 <- fct_explicit_na(f1)
# now
f2 <- fct_na_value_to_level(f1)

fct_count(f2)
table(f2)
sum(is.na(f2))
}
\keyword{internal}
