% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump.R
\name{fct_lump}
\alias{fct_lump}
\title{Superseded helper for lumping factor levels}
\usage{
fct_lump(
  f,
  n,
  prop,
  w = NULL,
  other_level = "Other",
  ties.method = c("min", "average", "first", "last", "random", "max")
)
}
\arguments{
\item{f}{A factor (or character vector).}

\item{n}{Positive \code{n} preserves the most common \code{n} values.
Negative \code{n} preserves the least common \code{-n} values.
If there are ties, you will get at least \code{abs(n)} values.}

\item{prop}{Positive \code{prop} lumps values which do not appear at least
\code{prop} of the time. Negative \code{prop} lumps values that
do not appear at most \code{-prop} of the time.}

\item{w}{An optional numeric vector giving weights for frequency of
each value (not level) in \code{f}.}

\item{other_level}{Value of level used for "other" values. Always
placed at end of levels.}

\item{ties.method}{A character string specifying how ties are
treated. See \code{\link[=rank]{rank()}} for details.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{fct_lump()} automatically chooses between \code{\link[=fct_lump_min]{fct_lump_min()}},
\code{\link[=fct_lump_prop]{fct_lump_prop()}}, \code{\link[=fct_lump_n]{fct_lump_n()}}, and \code{\link[=fct_lump_lowfreq]{fct_lump_lowfreq()}} based on
its arguments. It is kept for backward compatibility, but is
superseded and no longer recommended for new code.

For new code, prefer the more explicit helpers:
\code{\link[=fct_lump_min]{fct_lump_min()}}, \code{\link[=fct_lump_prop]{fct_lump_prop()}}, \code{\link[=fct_lump_n]{fct_lump_n()}}, and
\code{\link[=fct_lump_lowfreq]{fct_lump_lowfreq()}}.
}
\examples{
x <- factor(letters[rpois(100, 5)])
table(x)
table(fct_lump(x, n = 3))
}
