% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump.R
\name{fct_lump_helpers}
\alias{fct_lump_helpers}
\alias{fct_lump_min}
\alias{fct_lump_prop}
\alias{fct_lump_n}
\alias{fct_lump_lowfreq}
\title{Lump uncommon factor levels together into "other"}
\usage{
fct_lump_min(f, min, w = NULL, other_level = "Other")

fct_lump_prop(f, prop, w = NULL, other_level = "Other")

fct_lump_n(
  f,
  n,
  w = NULL,
  other_level = "Other",
  ties.method = c("min", "average", "first", "last", "random", "max")
)

fct_lump_lowfreq(f, w = NULL, other_level = "Other")
}
\arguments{
\item{f}{A factor (or character vector).}

\item{min}{Preserve levels that appear at least \code{min} number of times.}

\item{w}{An optional numeric vector giving weights for frequency of
each value (not level) in \code{f}.}

\item{other_level}{Value of level used for "other" values. Always
placed at end of levels.}

\item{prop}{Positive \code{prop} lumps values which do not appear at least
\code{prop} of the time. Negative \code{prop} lumps values that
do not appear at most \code{-prop} of the time.}

\item{n}{Positive \code{n} preserves the most common \code{n} values.
Negative \code{n} preserves the least common \code{-n} values.
If there are ties, you will get at least \code{abs(n)} values.}

\item{ties.method}{A character string specifying how ties are
treated. See \code{\link[=rank]{rank()}} for details.}
}
\description{
A family for lumping together levels that meet some criteria.
\itemize{
\item \code{fct_lump_min()}: lumps levels that appear fewer than \code{min} times.
\item \code{fct_lump_prop()}: lumps levels that appear in fewer than (or equal to)
\code{prop * n} times.
\item \code{fct_lump_n()} lumps all levels except for the \code{n} most frequent
(or least frequent if \code{n < 0})
\item \code{fct_lump_lowfreq()} lumps together the least frequent levels, ensuring
that "other" is still the smallest level.
}
}
\examples{
x <- factor(rep(LETTERS[1:9], times = c(40, 10, 5, 27, 1, 1, 1, 1, 1)))
x |> table()
x |>
  fct_lump_n(3) |>
  table()
x |>
  fct_lump_prop(0.10) |>
  table()
x |>
  fct_lump_min(5) |>
  table()
x |>
  fct_lump_lowfreq() |>
  table()
}
\seealso{
\code{\link[=fct_other]{fct_other()}} to convert specified levels to other.
}
