% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump.R
\name{fct_lump_min}
\alias{fct_lump_min}
\title{Title}
\usage{
fct_lump_min(f, min, w = NULL, other_level = "Other")
}
\arguments{
\item{f}{A factor (or character vector).}

\item{min}{Preserves values that appear at least \code{min} number of times.}

\item{w}{An optional numeric vector giving weights for frequency of
each value (not level) in f.}

\item{other_level}{Value of level used for "other" values. Always
placed at end of levels.}
}
\description{
Title
}
\examples{
x <- factor(letters[rpois(100, 5)])
fct_lump_min(x, min = 10)
}
