% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{fct_paste}
\alias{fct_paste}
\title{"Paste" together factors.}
\usage{
fct_paste(..., sep = " ")
}
\arguments{
\item{...}{Factors to be united.}

\item{sep}{Separtor to use}
}
\description{
Order is cartesian product of all levels.
}
\details{
This is sometimes useful when plotting a factor
}
\examples{
f <- factor(c("a", "b", "c"))
g <- factor(c("x", "y", "z"))
fct_paste(f, g, sep="-")
}

