% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered.R
\name{fct_remove_ordered}
\alias{fct_remove_ordered}
\alias{fct_make_ordered}
\title{Add/Remove the ordered attribute to factors.}
\usage{
fct_remove_ordered(f)

fct_make_ordered(f)
}
\arguments{
\item{f}{A factor}
}
\description{
Add/Remove the ordered attribute to factors.
}
\examples{
x <- factor(c("Gold", "Bronze", "Silver", "Silver", "Gold", "Gold"),
                     levels = c("Bronze", "Silver", "Gold"))
ordered_x <- fct_make_ordered(x)
ordered_x
fct_remove_ordered(ordered_x)
}
