% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{fct_replace}
\alias{fct_replace}
\alias{fct_replace_all}
\title{Transform levels of a factor with a regular expression}
\usage{
fct_replace(f, pattern, replacement)

fct_replace_all(f, pattern, replacement)
}
\arguments{
\item{f}{A factor}

\item{pattern, replacement}{Pattern and replacement regular expressions.
See \code{\link[stringr]{str_replace}}.}
}
\value{
A factor vector with the values of \code{f} and transformed levels.
}
\description{
Replace the factor levels with a regular expression.
\code{fct_replace} replaces the factor levels using
\code{\link[stringr]{str_replace}}, while \code{fct_replace_all} uses
\code{\link[stringr]{str_replace_all}}
}
\examples{
x <- factor(c("apple", "bear", "banana", "deer"))
fct_replace(x, "[aeiou]+", "-")
fct_replace_all(x, "[aeiou]+", "-")
}

