% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{fct_rev}
\alias{fct_rev}
\title{Reverse the levels of a factor}
\usage{
fct_rev(f)
}
\arguments{
\item{f}{A factor}
}
\description{
This is sometimes useful when plotting a factor
}
\examples{
f <- factor(c("a", "b", "c"))
fct_rev(f)
}

