% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{fct_shift}
\alias{fct_shift}
\title{Shift the order of levels of a factor}
\usage{
fct_shift(f, left_by = 1L)
}
\arguments{
\item{f}{A factor.}

\item{left_by}{A numeric indicating the amount that the levels will be
shifted left by. Negative values shift the levels to the right.}
}
\description{
This is useful when the levels of an ordered factor are actually cyclical,
with different conventions on the starting point.
}
\examples{
x <- factor(weekdays(Sys.Date()), levels = c("Sunday", "Monday", "Tuesday",
            "Wednesday", "Thursday", "Friday", "Saturday"), ordered = TRUE)
x
fct_shift(x)
fct_shift(x, left_by = 1)    # The same
}

