% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{fct_shuffle}
\alias{fct_shuffle}
\title{Randomly permute the levels of a factor}
\usage{
fct_shuffle(f)
}
\arguments{
\item{f}{A factor}
}
\description{
Randomly permute the levels of a factor
}
\examples{
f <- factor(c("a", "b", "c"))
fct_shuffle(f)
fct_shuffle(f)
}

