% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_levels.R
\name{fct_sort_levels}
\alias{fct_sort_levels}
\title{Automatically sort factor levels according to a user-defined function}
\usage{
fct_sort_levels(.f, .fun, ...)
}
\arguments{
\item{.f}{A factor (or character vector).}

\item{.fun}{A function that will sort or permute the existing factor levels.
It must accept one character argument and return a character argument of
the same length as it's input.}

\item{...}{Additional arguments to \code{.fun}.}
}
\description{
Automatically sort factor levels according to a user-defined function
}
\examples{
chromosomes <- c("chr2", "chr1", "chr10")
chr_fac <- factor(chromosomes, levels = chromosomes)
chr_fac

# naive alphanumeric sorting "1" < "10" < "2"
fct_sort_levels(chr_fac, sort)

 # number-based alphanumeric sorting "1" < "2" < "10"
 parse_number <- function(x){
   as.numeric(gsub(".*?([[:digit:]]+).*", "\\\\1", x))
 }
 sort_numeric <- function(x){
   x[order(parse_number(x))]
 }
 fct_sort_levels(chr_fac, sort_numeric)

}
