% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{fct_unify}
\alias{fct_unify}
\title{Unify the levels in a list of factors}
\usage{
fct_unify(fs, levels = lvls_union(fs))
}
\arguments{
\item{fs}{A list of factors}

\item{levels}{Set of levels to apply to every factor. Default to
union of all factor levels}
}
\description{
Unify the levels in a list of factors
}
\examples{
fs <- list(factor("a"), factor("b"), factor(c("a", "b")))
fct_unify(fs)
}
