% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{fct_unique}
\alias{fct_unique}
\title{Unique values of a factor}
\usage{
fct_unique(f)
}
\arguments{
\item{f}{A factor}
}
\description{
Unique values of a factor
}
\examples{
f <- factor(letters[rpois(100, 10)])

unique(f)     # in order of appearance
fct_unique(f) # in order of levels
}
