% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcats-package.R
\docType{package}
\name{forcats-package}
\alias{forcats}
\alias{forcats-package}
\title{forcats: Tools for Working with Categorical Variables (Factors)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helpers for reordering factor levels (including moving specified levels to front, ordering by first appearance, reversing, and randomly shuffling), and tools for modifying factor levels (including collapsing rare levels into other, 'anonymising', and manually 'recoding').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forcats.tidyverse.org/}
  \item \url{https://github.com/tidyverse/forcats}
  \item Report bugs at \url{https://github.com/tidyverse/forcats/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
