% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{lvl_union}
\alias{lvl_union}
\title{Find all levels in a list of factors}
\usage{
lvl_union(fs)
}
\arguments{
\item{fs}{A list of factors.}
}
\description{
Find all levels in a list of factors
}
\examples{
fs <- list(factor("a"), factor("b"), factor(c("a", "b")))
lvl_union(fs)
}

