% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvls.R
\name{lvls_union}
\alias{lvls_union}
\title{Find all levels in a list of factors}
\usage{
lvls_union(fs)
}
\arguments{
\item{fs}{A list of factors.}
}
\description{
Find all levels in a list of factors
}
\examples{
fs <- list(factor("a"), factor("b"), factor(c("a", "b")))
lvls_union(fs)
}

