% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{fct_collapse}
\alias{fct_collapse}
\title{Collapse rare levels into "other".}
\usage{
fct_collapse(x, threshold = 0.1, other_level = "Other")
}
\arguments{
\item{x}{A factor.}

\item{threshold}{Values with less than this proportion of entries will
be collpased}

\item{other_level}{Value of level used for "other" values.}
}
\description{
Collapse rare levels into "other".
}
\examples{
x <- factor(letters[rpois(100, 5)])
x
fct_collapse(x)
}

