% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{fct_collapse}
\alias{fct_collapse}
\title{Collapse rare levels into "other".}
\usage{
fct_collapse(x, n, prop, other_level = "Other")
}
\arguments{
\item{x}{A factor.}

\item{n, prop}{Either specify \code{n} or \code{prop}.

  If you specify \code{n}, only the most common \code{n} values will be kept
  (unless there are ties, in which case it will keep more).

  If you specify \code{prop}, only values that appear at least that
  commonly will be kept.}

\item{other_level}{Value of level used for "other" values.}
}
\description{
Collapse rare levels into "other".
}
\examples{
x <- factor(letters[rpois(100, 5)])
x
fct_collapse(x, n = 3)
fct_collapse(x, prop = 0.1)
}

