% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{fct_collapse}
\alias{fct_collapse}
\title{Collapse factors into groups.}
\usage{
fct_collapse(f, ...)
}
\arguments{
\item{f}{A factor}

\item{...}{A series of named character vectors. The levels in
each vector will be replaced with the name.}
}
\description{
Collapse factors into groups.
}
\examples{
fct_collapse(gss_cat$partyid,
  missing = c("No answer", "Don't know"),
  other = "Other party",
  rep = c("Strong republican", "Not str republican"),
  ind = c("Ind,near rep", "Independent", "Ind,near dem"),
  dem = c("Not str democrat", "Strong democrat")
)
}

