% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{fct_expand}
\alias{fct_expand}
\title{Add additional levels to a factor}
\usage{
fct_expand(f, ...)
}
\arguments{
\item{f}{A factor.}

\item{...}{Additional levels to add to the factor.  Levels that already
exist will be silently ignored.}
}
\description{
Add additional levels to a factor
}
\examples{
f <- factor(sample(letters[1:3], 20, replace = TRUE))
f
fct_expand(f, "d", "e", "f")
fct_expand(f, letters[1:6])
}

