% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inorder.R
\name{fct_inorder}
\alias{fct_inorder}
\title{Reorders levels in order of first appearance in data}
\usage{
fct_inorder(f)
}
\arguments{
\item{f}{As factor}
}
\description{
Reorders levels in order of first appearance in data
}
\examples{
f <- factor(c("b", "b", "a", "a", "c", "c"))
f
fct_inorder(f)
}

