% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{fct_inorder}
\alias{fct_inorder}
\alias{fct_infreq}
\alias{fct_inseq}
\title{Reorder factors levels by first appearance, frequency, or numeric order.}
\usage{
fct_inorder(f, ordered = NA)

fct_infreq(f, ordered = NA)

fct_inseq(f, ordered = NA)
}
\arguments{
\item{f}{A factor}

\item{ordered}{A logical which determines the "ordered" status of the
output factor. \code{NA} preserves the existing status of the factor.}
}
\description{
Reorder factors levels by first appearance, frequency, or numeric order.
}
\examples{
f <- factor(c("b", "b", "a", "c", "c", "c"))
f
fct_inorder(f)
fct_infreq(f)

fct_inorder(f, ordered = TRUE)

f <- factor(sample(1:10))
fct_inseq(f)
}
