% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{fct_match}
\alias{fct_match}
\alias{fct_exclude}
\title{Test for presence of certain groups/levels in a vector}
\usage{
fct_match(f, ..., inverse = FALSE, allow_missing = TRUE)

fct_exclude(f, ..., allow_missing = TRUE)
}
\arguments{
\item{f}{A factor (or character vector).}

\item{...}{A series of characters for which the test "are these levels in
\code{f}?" should be performed.}

\item{inverse}{(logical) if \code{TRUE}, perform a "not match" operation.}

\item{allow_missing}{(logical) if \code{TRUE}, allow matching to empty levels in
f}
}
\value{
a logical vector determined by whether or not the levels in \code{...} are
present in \code{f}
}
\description{
A validated version of \code{\%in\%} intended for use with logical subsetting.
Performs matching between a factor and a set of levels to pose the question
"which of the elements of a factor match these levels?"
}
\details{
Unlike \code{\%in\%}, which performs a basic yes/no test, this function
performs validation on the input vector to ensure that the levels tested
for are all present before creating a logical vector. If a tested level is
not present, then an error is thrown rather than producing  a silent
\code{FALSE}.

Since a distinction can be made between \code{levels(f)} and \code{unique(f)} (the
latter excludes empty levels) the argument \code{allow_missing} can be
specified.

\code{fct_exclude} is a shortcut to \code{fct_match(inverse = TRUE)}.

This is intended to replace the common pattern

\code{filter(d, l \%in\% c("x", "y"))}

with the validated

\code{filter(d, fct_match(l, c("x", "y")))}

where an error will be generated if \code{"x"} or \code{"y"} are not levels in \code{d$l}.
}
\examples{
## TRUE if present
gss_cat$marital \%>\% fct_match("Married", "Divorced")
## TRUE if not present (exclude)
gss_cat$marital \%>\% fct_match("No answer", inverse = TRUE)
gss_cat$marital \%>\% fct_exclude("No answer")
## importantly, misspelled levels throw an error
\dontrun{gss_cat$marital \%>\% fct_match("Married", "Davorced")}
## match to levels()
gss_cat$race \%>\% fct_match("Other", "Not applicable")
## match to unique()
gss_cat$race \%>\% fct_match("Other", "Not applicable", allow_missing = TRUE)
}
