% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{fct_reorder}
\alias{fct_reorder}
\title{Reorder the levels of a function according to another variable}
\usage{
fct_reorder(x, y, fun = median, ...)
}
\arguments{
\item{x}{A factor}

\item{y, fun}{fun The levels of \code{x} will be reordered so that \code{fun}
apply to each group is in ascending order.}

\item{...}{Other arguments passed on to \code{fun}. A common argument is
\code{na.rm = TRUE}.}
}
\description{
\code{fct_reorder} is useful for 1d displays where the factor is mapped to
position; \code{fct_reorder2} for 2d displays where the factor is mapped to
a non-position aesthetic.
}
\details{
\code{fct_reorder} is currently an almost exact copy of
\code{\link[stats]{reorder}} but it may change in the future.
}
\examples{
boxplot(Sepal.Width ~ Species, data = iris)
boxplot(Sepal.Width ~ fct_reorder(Species, Sepal.Width), data = iris)
}

