% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{fct_shuffle}
\alias{fct_shuffle}
\title{Randomly permute factor levels}
\usage{
fct_shuffle(f)
}
\arguments{
\item{f}{A factor (or character vector).}
}
\description{
Randomly permute factor levels
}
\examples{
f <- factor(c("a", "b", "c"))
fct_shuffle(f)
fct_shuffle(f)
}
