% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_levels.R
\name{fct_sort_levels}
\alias{fct_sort_levels}
\title{Sort factor levels according to a user-defined function}
\usage{
fct_sort_levels(.f, .fun, ...)
}
\arguments{
\item{.f}{A factor (or character vector).}

\item{.fun}{A function that will sort or permute the existing factor levels.
It must accept one character argument and return a character argument of
the same length as it's input.}

\item{...}{Additional arguments to \code{.fun}.}
}
\description{
Sort factor levels according to a user-defined function
}
\examples{
medieval_experiment <- factor(
  c("Control", "Comfrey", "Cropleek", "Bishopwort")
)
medieval_experiment

# Default (alphabetical) level-sorting:
fct_sort_levels(medieval_experiment, sort)

# Reversed ordering (equivalent to `fct_rev`):
fct_sort_levels(medieval_experiment, sort, decreasing = TRUE)

# Level-sorting with "Control" as the first level
# (equivalent: `fct_relevel(medieval_experiment, "Control")`)
fct_sort_levels(
  medieval_experiment,
  function(x) c("Control", sort(setdiff(x, "Control"))))

# Randomised sorting of the levels:
fct_sort_levels(medieval_experiment, sample)

}
