% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcats-package.r
\docType{package}
\name{forcats}
\alias{forcats}
\alias{forcats-package}
\title{forcats.}
\description{
Helpers for reordering factor levels (including moving specified levels to
front, ordering by first appearance, reversing, and randomly shuffling), and
tools for modifying factor levels (including collapsing rare levels into
other, anonymising, and manually recoding).
}

